Make the CDI/interceptor dependencies optional in module-info
The "static" keyword means "required for compilation of this module, but
not at runtime", which is pretty much the current situation.

Those dependencies are only required for the @Transactional
and @TransactionScoped annotations,
but jakarta.transaction can perfectly well be used without CDI, e.g. in
Hibernate ORM in a Java SE environment.

Without this change, it's impossible to use jakarta.transaction-api (or
anything relying on it, e.g. Hibernate ORM) without CDI in the
modulepath.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>